Faster! More Reliable!
Decided to switch over to using direct registry changes for the color theming, and using the win32 API for setting the wallpaper. This cuts down on loading time.

Additionally, I found an exact method, without using magic numbers, by using the clipboard as a temporary register for marking theme changes as done.

Also, I was able to find different window states that work better with preventing white and blank windows.

All in all, now it loads faster, and is on point on the timing for restoring the window (no more guessing)